Updated: change location of promptFormat
